% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosonora-package.R
\docType{package}
\name{biosonora-package}
\alias{biosonora}
\alias{biosonora-package}
\title{biosonora: Acoustic Biodiversity Monitoring with BirdNET and Observation.org Export}
\description{
Shiny application to organise, visualise, listen to, review and export field sound recordings (AudioMoth, TASCAM and similar), integrating BirdNET detections and exporting validated observations to Observation.org.
}
\author{
\strong{Maintainer}: Antonio Canepa \email{ajcanepa@gmail.com}

Authors:
\itemize{
  \item Antonio Canepa \email{ajcanepa@gmail.com}
}

}
\keyword{internal}
